AtualizacaoJWT

+++ src/controllers/reservaController.js
@@ -1,5 +1,6 @@
 const validateReserva = require("../services/validateReserva");
 const { queryAsync, formatarHorario } = require("../utils/functions");
+const jwt = require("jsonwebtoken");
 
 // Retorna o dia da semana em português, dado uma data no formato "YYYY-MM-DD"
 const getDiaSemana = (data) => {
@@ -20,6 +21,7 @@ const getDiaSemana = (data) => {
 module.exports = class ReservaController {
   static async createReservas(req, res) {
     const { fk_id_usuario, fk_id_sala, data, hora_inicio, hora_fim } = req.body;
+    const verificarToken = req.userId;
 
     // Validação inicial dos campos
     const erroValidacao = validateReserva.validarCamposReserva({
@@ -31,6 +33,9 @@ module.exports = class ReservaController {
     });
     if (erroValidacao) {
       return res.status(400).json(erroValidacao);
+    }
+    if (Number(verificarToken) !== Number(fk_id_usuario)){
+      return res.status(400).json({message: "Você não pode fazer reserva com ID de outro usuário!"});
     }
 
     try {
@@ -104,17 +109,23 @@ module.exports = class ReservaController {
   }
 
   static async updateReserva(req, res) {
-    const { data, hora_inicio, hora_fim } = req.body;
+    const { fk_id_usuario, data, hora_inicio, hora_fim } = req.body;
     const reservaId = req.params.id_reserva;
+    const verificarToken =  req.userId;
 
     // Valida os campos de atualização
     const erroValidacao = validateReserva.validarCamposAtualizacao({
+      fk_id_usuario,
       data,
       hora_inicio,
       hora_fim,
     });
     if (erroValidacao) {
       return res.status(400).json(erroValidacao);
+    }
+
+    if (Number(verificarToken) !== Number(fk_id_usuario)){
+      return res.status(400).json({message: "Você não pode atualizar a reserva de outro usuário!"});
     }
 
     try {
@@ -165,6 +176,12 @@ module.exports = class ReservaController {
 
   static async deleteReserva(req, res) {
     const reservaId = req.params.id_reserva;
+    const usuarioId = req.params.id_usuario;
+    const verificarToken =  req.userId;
+
+    if (Number(verificarToken) !== Number(usuarioId)){
+      return res.status(400).json({message: "Você não pode deletar a reserva de outro usuário!"});
+    }
     const query = `DELETE FROM reserva WHERE id_reserva = ?`;
     try {
       const results = await queryAsync(query, [reservaId]);
